Revert "Feature/app structure"
