--- frontend/src/app/home/components/HomeFooter.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-
-import { CTASection } from '../../../components/CTASection';
-import { Footer } from '../../../components/Footer';
-
-/**
- * Home page footer section component.
- * Pure UI component containing CTA and footer sections.
- */
-export function HomeFooter() {
-  return (
-    <>
-      <CTASection />
-      <Footer />
-    </>
-  );
-}
