--- frontend/src/app/home/components/HomeHero.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import { HeroSection } from '../../../components/HeroSection';
-
-/**
- * Home page hero section component.
- * Pure UI component wrapping the main hero section.
- */
-export function HomeHero() {
-  return <HeroSection />;
-}
