--- frontend/src/app/home/components/index.ts
@@ -1,4 +1,0 @@
-export { HomeFooter } from './HomeFooter';
-export { HomeHeader } from './HomeHeader';
-export { HomeHero } from './HomeHero';
-export { HomeMainContent } from './HomeMainContent';
