--- frontend/src/app/home/hooks/useHomePage.ts
@@ -1,18 +1,0 @@
-/**
- * Custom hook for home page logic and state management.
- * Separates business logic from UI presentation.
- * Currently minimal but provides structure for future features.
- * 
- * @returns Object containing home page state and handlers
- */
-export function useHomePage() {
-  // Future home page logic can go here:
-  // - Analytics tracking
-  // - User preferences
-  // - Dynamic content loading
-  // - Navigation state
-  
-  return {
-    // Placeholder for future state
-  };
-}
