--- frontend/src/app/login/components/index.ts
@@ -1,5 +1,0 @@
-export { LoginFormFields } from './LoginFormFields';
-export { LoginHeader } from './LoginHeader';
-export { LoginMessages } from './LoginMessages';
-export { LoginSocialOptions } from './LoginSocialOptions';
-export { LoginSubmitButton } from './LoginSubmitButton';
