--- frontend/src/app/login/page.tsx
@@ -1,12 +1,0 @@
-import { LoginForm } from './LoginForm';
-
-import type { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'Login - Language Exchange',
-  description: 'Sign in to your Language Exchange account',
-};
-
-export default function LoginPage() {
-    return <LoginForm />;
-}
