--- frontend/src/app/register/components/index.ts
@@ -1,2 +1,0 @@
-export { RegisterFormActions } from './RegisterFormActions';
-export { RegisterMessages } from './RegisterMessages';
