--- frontend/src/app/types/index.ts
@@ -1,5 +1,0 @@
-/**
- * App Types - Barrel Export
- * Central export for all app-specific types
- */
-export type { ErrorComponentProps, TemplateProps } from './next-components';
